backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
openai==1.82.0
pydantic==2.10.4
python-multipart==0.0.20
httpx==0.28.1
supabase==2.13.0
psycopg2-binary==2.9.10
stripe>=8.0.0
